backend/apps/groups/serializers.py
from .models import Group
from rest_framework import generics
from rest_framework import serializers


class GroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = Group
        fields = ('id', 'name', 'description')
